+++ ft_striteri.c
@@ -6,7 +6,7 @@
 /*   By: kirus <kirus@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/19 21:58:30 by kirus             #+#    #+#             */
-/*   Updated: 2021/10/25 03:13:19 by kirus            ###   ########.fr       */
+/*   Updated: 2021/10/29 20:45:35 by kirus            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char*))
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (s == NULL || f == NULL)
