+++ ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: kirus <kirus@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/19 21:36:56 by kirus             #+#    #+#             */
-/*   Updated: 2021/10/24 03:01:42 by kirus            ###   ########.fr       */
+/*   Updated: 2021/10/29 20:46:21 by kirus            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
 	char	*str;
-	int		i;
+	size_t	i;
 
 	i = 0;
 	if (s == NULL || f == NULL)
