+++ ft_strrchr.c
@@ -14,8 +14,8 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	int				i;
-	int				p;
+	size_t			i;
+	size_t			p;
 	char			*str;
 	unsigned char	ch;
